run_make_support: move `run_in_tmpdir` and `test_while_readonly` to `scoped_run` module
